+++ CHANGELOG.md
@@ -5,7 +5,14 @@ All notable changes to this project are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.0.2] - 2026-08-27
+
+### Changed
+
+- Echo the VirtFusion API token while pasting
+- Strip trailing `/` from Proxmox, VirtFusion, and SolusVM API URLs
+- Issue Mimir write tokens via API instead of prompting
+- Fail register if `--agent-ip` is not on a local interface
 
 ## [1.0.1] - 2026-08-27